Fix application --> daemon communication
This was preventing the authentication window from coming up when enforcing.

We add a new device type of su_device, being an mlstrustedobject so that untrusted_app can communicate with a socket labeled su_device. We establish a file_type_auto_trans from device to su_device for su context, and we grant access to those su_device files to untrusted_app.