Update prebuilt boot images
For Angler, Bullhead, Shamu. Note that AOSP source may not be
up to date with final release for Angler or Shamu -- as of this
time, neither has available final factory images of 7.0.

However, *probably* they will both work just fine, since there
are builds of 7.0r1 based on AOSP source running on Shamu.