Update for 7.1.1_r28
Also note that the kernel in this has been patched to apply PWM to
the LED. I wanted to enable the blinker, but we are denied the
needed firmware for that.

The qcom LPG (Light Pattern Generator) has a number of different
modes. Basic mode is disabled (leaving the LEDs in a "GPIO on/off"
state), and PWM (pulse width modulation), which is just brightness.

In addition, there are two USER MODES available, which can be
programmed with special patterns based on what the customer
(google, moto) ask for. These include special functions like
"blink" and "breathe". Unfortunately for us, they just left them
BLANK.